%i %d %u done with -s (everything) FUCK YEAH

+++ cursus/common_core_1/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: blvilarn <blvilarn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/09 19:33:01 by blvilarn          #+#    #+#             */
-/*   Updated: 2023/01/10 19:50:10 by blvilarn         ###   ########.fr       */
+/*   Updated: 2023/01/11 19:22:49 by blvilarn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ int	ft_printf(const char *str, ...);
 int	print_content(const char *str, va_list ptr, int *i, int *len);
 int	ft_putchar(char c, int *i, int *len);
 int	ft_putstr(char *s, int *i, int *len);
-int	ft_putnumber(int n, int *i, int *len);
+int	ft_putnumber(long long n, int *i, int *len);
 
 #endif
